backend/ml_server/requirements.txt
# ml_server/requirements.txt
fastapi
uvicorn
torch
torchvision
timm
albumentations
pillow
requests
python-dotenv